requirements.txt
amqp==5.0.5
asgiref==3.3.1
backcall==0.2.0
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
decorator==4.4.2
Django==3.1.6
django-debug-toolbar==3.2
django-extensions==3.1.1
Faker==6.5.2
flake8==3.8.4
flake8-builtins==1.5.3
flake8-print==4.0.0
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.0.2
mccabe==0.6.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.4
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.0.5
vine==5.0.0
wcwidth==0.2.5
xlwt==1.3.0
